+++ backend/src/main/java/com/fschoen/parlorplace/backend/enumeration/NotificationType.java
@@ -1,0 +1,6 @@
+package com.fschoen.parlorplace.backend.enumeration;
+
+public enum NotificationType {
+    STALE_GAME_INFORMATION,
+    VOICE_LINE
+}