+++ backend/src/main/java/com/fschoen/parlorplace/backend/enumeration/StaleType.java
@@ -1,5 +1,0 @@
-package com.fschoen.parlorplace.backend.enumeration;
-
-public enum StaleType {
-    GAME
-}