+++ frontend/src/environments/environment.ts
@@ -4,4 +4,5 @@ export const environment = {
   general: environmentGeneral,
   production: false,
   BASE_URI: "http://localhost:8080/",
+  WEBSOCKET_BASE_URI: "ws://localhost:8080/",
 };